+++ client/app/instance-initializers/current-user.js
@@ -15,7 +15,7 @@ export default {
 
     instance.registry.injection('controller', 'currentUser', 'user:current');
     instance.registry.injection('route',      'currentUser', 'user:current');
-    instance.registry.injection('component:overlay-task', 'currentUser', 'user:current');
+    instance.registry.injection('component', 'currentUser', 'user:current');
     instance.registry.injection('service:notifications', 'currentUser',  'user:current');
   }
 };