+++ config/initializers/subscriptions.rb
@@ -8,7 +8,7 @@ Subscriptions.configure do
   add '.*', EventLogger
   add 'paper:submitted', Paper::Submitted::EmailCreator, Paper::Submitted::EmailAdmins, Paper::Submitted::SnapshotMetadata
   add 'paper:initially_submitted', Paper::Submitted::SnapshotMetadata
-  add 'paper:resubmitted', Paper::Resubmitted::EmailEditor
+  add 'paper:resubmitted', Paper::Resubmitted::EmailAcademicEditors
 
   add 'discussion_reply:created', \
     DiscussionReply::Created::EmailPeopleMentioned, DiscussionReply::Created::NotifyPeopleMentioned