+++ app/services/epub_converter.rb
@@ -28,7 +28,6 @@ class EpubConverter
               locals: {
                 paper: paper,
                 paper_body: paper_body,
-                orphan_figures: orphan_figures,
                 title: title,
                 converter: self
               })