+++ app/services/pdf_converter.rb
@@ -21,7 +21,6 @@ class PDFConverter
               locals: {
                 paper: @paper,
                 paper_body: paper_body,
-                orphan_figures: orphan_figures,
                 publishing_info_presenter: @publishing_info,
                 converter: self
               })